Right Margin and Dual Folder Fix
Fixed issue with the extra right margin causing page to be off center.
There was a parent row encapsulating all of the other rows. This was
removed, because it was not needed.

This issue can now be closed.
https://github.com/harryrf/portfolio/issues/4

There was also an issue with there being two "portfolio" folders. Only
one is needed. I added the .git to the inside of the portfolio folder
that was left.