Comment form, about section and mobile nav fixed
-Comment form needed some more love when it came to sizing for each
breakpoint. The "send" button is sitll a little bit off, but it's okay
for now.

-About section pic and info are both fixed. Did some more work and honed
in more on the breakpoints. I also added a media inquiry to keep the
info text in the proper place.

-The nav bar was a simple fix, but hard to find. I had bootstrap's JS
script tags above the jQuery tags in the HTML file. After moving the
tags to the bottom of the JS include list, everything worked out fine.